Deleted data resiliency card
